passing tests for editing stations
