+++ app/views/admin/stations/edit.html.erb
@@ -1,0 +1,1 @@
+<%= render "form" %>
