+++ app/views/admin/stations/new.html.erb
@@ -1,0 +1,1 @@
+<%= render "form" %>
